refactor: Update .gitignore and view_youtube_wjspering.ejs
The .gitignore file has been updated to include the 'public/pdf_video_transcrips/*' directory. This change ensures that the 'public/pdf_video_transcrips' directory and its contents are ignored by Git.

In the view_youtube_wjspering.ejs file, the navbar has been updated to include a background image and a fixed height. This improves the visual appearance of the navbar and enhances the overall design of the webpage.