refactor: Import User model and add route for creating mockup user
This commit imports the User model in the dbRoutes.js file to enable the creation of a mockup user. It adds a new route '/mockup' that creates a new user document using a mockup user object and saves it to the database. This enhancement provides a way to easily add a mockup user for testing purposes.

Refactor the code to improve database functionality and testing capabilities.