refactor: Update forgot password email link to use HTTPS
This commit updates the forgot password email link in the user_routes.js file to use HTTPS instead of HTTP. By using HTTPS, the link ensures a secure connection for users when resetting their password. This enhancement improves the security of the application and protects user data.

Refactor the code to enhance the security of the forgot password feature.