refactor: Update image path in waitinglist.html
This commit updates the image path in the waitinglist.html file to use the correct capitalization for the image file name. It changes the image source from "../images/soundanalytics.png" to "../images/SoundAnalytics.png". This change ensures that the image is correctly displayed and improves the consistency of the codebase.

Refactor the code to fix image path capitalization in waitinglist.html.