refactor: Update webpage field in User model and registration link in waitinglist.html
Refactor the User model in User.js to update the webpage field from "WebPage" to "webpage" and make it optional. This change allows for storing and retrieving the webpage associated with a user.

Update the registration link in waitinglist.html to point to the "bodymind" page instead of the "waitinglist" page. This change ensures that users are directed to the correct page when registering.

Refactor the code to improve the User model and registration link handling.