refactor: Add GET route for /vegvisr in products_route.js
Add a new GET route for /vegvisr in the products_route.js file. This route retrieves all vegvisrs from the database and returns them as a JSON response. This change improves the functionality of the API by allowing users to fetch all vegvisrs easily.

Refactor the code to include a new GET route for /vegvisr in products_route.js.