refactor: Add /vegvisr route to server.js
Add a new route, '/vegvisr', to the server.js file. This route is responsible for handling requests related to the Vegvisr module. By adding this route, we improve the modularity and organization of our codebase.

Refactor the code to include a new route for the Vegvisr module in server.js.