refactor: Update menu.html to fix pricing page link
Update the menu.html file to fix the link to the pricing page. The current link is pointing to "#" which is incorrect. This change updates the link to point to "../pricing.html" to ensure that users are directed to the correct pricing page.

Refactor the code to fix the pricing page link in menu.html.