refactor: Update navigation menu to include SlowYou™ Calendar link
Update the navigation menu in menu.html to include a link to the SlowYou™ Calendar page. This change adds a new list item with the link and ensures that users can easily access the calendar feature.

Refactor the code to update the navigation menu and add a link to the SlowYou™ Calendar page.