refactor: Update contacts.html to add "Add Contact" functionality and form
Refactor the contacts.html file to add the "Add Contact" functionality and form. This change includes the necessary HTML, CSS, and JavaScript code to display the form when the "Add New Contact" button is clicked. Users can now easily add new contacts by filling out the form fields and submitting the form. This enhancement improves the usability and functionality of the contacts page.